content_based/requirements.txt
polars==0.18.6
tensorboard==2.13.0
tensorflow==2.13.0
pyarrow==12.0.1
tqdm==4.65.0
scipy==1.11.1
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
sentence_transformers